requirements-api.txt
prometheus-client>=0.23.1
structlog>=25.5.0
pandas>=2.3.3
numpy>=2.4.0
requests>=2.32.5
SQLAlchemy>=2.0.45
minio>=7.2.9
celery>=5.4.0
redis>=5.2.0
psycopg2-binary>=2.9.9
confluent-kafka>=2.3.0
pytest-asyncio>=0.23.0
authlib>=1.3.0
httpx>=0.27.0
python-jose>=3.3.0
fastavro>=1.9.0
uvloop>=0.19.0
uvicorn>=0.29.0
fastapi>=0.109.1
orjson>=3.9.0
python-dotenv>=1.0.0
strawberry-graphql[federation]~=0.200
scipy>=1.15.0
cachetools>=5.5.0
alembic>=1.13.0
pgvector>=0.2.0
jinja2>=3.1.3
aiohttp>=3.9.4
